Merged PR 31: Implement a Manipulator component with hand rotation closely matching the Unity MRTK feature set
`GrabbableComponent`
Component base class to detect and manage grabbing pointers on interactable components. Grab pointers are inserted in an array, so that they remain sorted by start time and "primary" and "secondary" grabs can be determined regardless of handedness. An event dispatcher is provided for begin/end of grabs.

`Manipulator`
Blueprint class based on GrabbableComponent that serves as a reference implementation for a manipulation handler. It implements one-handed rotation modes similar to the Unity-MRTK ObjectManipulator component.

`ManipulationHandlerFunctionLibrary`
C++ function library can be utilized in blueprints for common features based on pointer input, such as:
* Finding world-space transforms of both the grab point (where the object was pinched) as well as the target point (where the pointer is now)
* Computing offsets and rotations based on a pointer
* Finding the centroid of all currently grabbing pointers, for two-hand interactions
* Future additions could be computing rotations from multiple pointers (requires solving an Eigenvalue problem, TBD) and support for more than two hands, e.g. for multi-user interactions (two people carrying a box)

`InteractableSceneComponent` renamed to `InteractableComponent`.

`TouchPointer` has an event dispatcher now for pinch begin/end, instead of the functions in TouchPointerTarget.

Demos/ObjectManipulator/ManipulatorTest is a test level for the Manipulator component.

Related work items: #84