Merged PR 74: Turn the LoadAllMaps test into a "complex" test with a parameter for each map
Turn the LoadAllMaps test into a "complex" test with a parameter for each map found.

This will load each map in a clean game instance, rather than replacing the currently loaded map. The UE4 docs list this as the example case for using a complex test:
https://docs.unrealengine.com/en-US/Programming/Automation/TechnicalGuide/#complextests

Note that this will require restarting the editor when maps are added/removed in order to update the test list.

Related work items: #131