Merged PR 73: Optional background target object for touch pointers
This works similar to a "canvas" in classic 2D UI. If no target is hovered then the hover and grasp events will be called on the background object instead. This allows handling grasps in empty space and detect when the pointer stops hovering actual objects.

New tests have been added for the background object. The pointer test utils now also support grasp events.