Merged PR 76: Optimize default ticking behavior of the UGrabbableComponent by only ticking while grabbed.
Optimize default ticking behavior of the UGrabbableComponent by only ticking while grabbed.

The bTickOnlyWhileGrabbed flag is enabled by default and controls the component tick. When a pointer starts or stops grasping, and this flag is enabled, the component tick function will be enabled or disabled. If the flag is not enabled the component tick is left untouched and can be modified elsewhere.

A new test has been added to confirm expected behavior of the tick flag.

Also set the game mode for the TestEmpty map to GameModeBase to avoid creating additional FingerPointer instances through our MRPawn.