Merged PR 97: Fix follow component getting stuck behind player
See #174 for more details, but it was possible, if you spun around quick enough, for an actor with the follow component to get stuck behind you. I think what was happening is that if the angle difference between `CurrentRefForward` and `ToTarget` was significant enough (⪆ 90° iirc), corrective rotations were applied twice (once for each projection we were measuring in) and so the new to target orientation would end up behind you again. This would resonate and find equilibrium behind you so that you would have to spin around quickly again, in order to get the component unstuck.

A fix that worked in MRTK (although there is a slight tweak made here, which I will describe) was to calculate the angle from the `ToTarget` vector and reference xy and xz planes and use that for corrective rotations.

Well this made it harder for the problem to occur, but there was still a sweet spot where if the actor ended up in the inverted follow frustum behind them, it would be stuck again.

The fix for this problem was to use the angle from `CurrentRefForward` and `ToTarget` projected onto the reference xz plane for vertical leashing, but use the angle from `ToTarget` to the reference xz plane for horizontal leasing. Ensuring that the inverted frustum is not a valid leashing position.

I'll make this fix for MRTK now as well.

Related work items: #174