Merged PR 102: Input simulation assets for substituting device data at runtime.
The InputSimulationGameInstanceSubsystem is created when a game starts. It creates an actor in the world that manages input, movement, and animations for the camera and hands. The resulting actor and bone transforms are copied to the WMR input sim subsystem as the HMD pose and hand joint transforms.

The input sim actor binds to user input, by which it can be moved with WASD keys, as well as Q/E keys for vertical movement (this is based on the UFloatingPawnMovement). By holding the Left-Shift and Left-Alt keys the control of left/right hands can be enabled instead of rotating the camera.

The actor also keeps a stack of active hand poses that can be enabled with mouse clicks (Left = Pinch, Middle = Poke). The minutiae of the hand animation are handled by an animation blueprint (see the InputSimulation.md document for details).

A new settings module has been added to store configuration. This contains default settings for the input simulation but can be changed by users if desired.

Related work items: #83