Merged PR 111: Updated stale references in pointer highlight material
Updated stale references to pointer highlight function and pointer positions parameter collection in the pointer highlight material.

It seems that the renaming from MRU to UXT did not update correctly these references, even though they load and work fine in the editor when working with plugin sources. This only shows when loading or trying to package the game using a prebuilt plugin with errors like:
DoesPackageExist FAILED: '/MixedRealityUtils/FingerPointer/PointerHighlightFunction' is not a standard unreal filename or a long path name

Related work items: #210