Merged PR 128: Fix pointer tests to ensure expected conditions are actually checked
The conditional for testing pointer movement keyframes was never met, so these tests were not reliably running.
The tests also have to give the pointer one additional frame to update overlaps to make sure the focused/grabbed object are correct at the time of testing.

The default (fallback) target for touchables has been removed because it wasn't very useful and adds a number of complex corner cases wrt. to raising events on the first frame, and deciding what happens when changing the default target.

Related work items: #234