Merged PR 176: Merged touch target component functionality into button component one.
Merged touch target component functionality into button component one.

We don't have any other touch control at the moment and the button component needs to be modified to be also a far target so its makes sense to fold the functionality into the button to keep things simple and the public API small. To be revisited once we have more touch-enabled controls.