Merged PR 190: Make sure that changes to button's public properties take effect at runtime
Make sure that the public properties of the button can be configured at runtime.

Also includes some changes for the push distance calculation so that it is always measured from the front face rather than from the transform location. This makes it easier for users to create buttons in the editor, but also at runtime, as the transform location does not have to align with the front face of the mesh.

Related work items: #282