Merged PR 228: Pressable Button Tests
Adds a number of test cases for PressableButton. Most of these tests ensure that the correct button events are called upon pressing and releasing the button in various ways.

There is also a test that ensures that a button does not cause errors if it is destroyed during a press.

There are currently no two handed tests.

Related work items: #286