Merged PR 259: Fix issue where cursor would get stuck in certain circumstances
The finger cursor was failing to update it's position under certain conditions. We have a distance check against an epsilon that was intended to prevent inconsistent rotations, but at some stage this was erroneously changed to include position too. As a result, the cursor would sometimes get stuck:

This change continues to update the cursor position even when the distance to the target is smaller than the epsilon.

Related work items: #319