Merged PR 324: Remove the old PointerTestSequence in favor of direct FrameQueue usage.
Remove the old PointerTestSequence in favor of direct FrameQueue usage.

The PointerTestSequence was a way of organizing pointer focus and grab tests.
It has been superceeded by the use of automation test specs and the FrameQueue,
which essentially serves the same purpose. The NearPointerComponent tests are
the only tests still using the PointerTestSequence, so tests can be simplified
a lot by using a FrameQueue directly.

NearPointerComponent tests have been combined in a single file.

Related work items: #419