Merged PR 328: Fix near interaction starting when pinching before focusing the target
This was caused by the near pointer only checking for the current state of the hand for it's grab logic.
As you moved close to an object while pinching, as the near pointer activates it will grab the nearest object as it detects the pinch.

The fix is to track the state of the hand for the previous frame and only grab when the pinch begins while the near pointer is active.

Related work items: #401