Merged PR 344: UxtIconBrush Type and Editor Window
Before this change icon selection was a very manual process of looking up the icon you wanted in an external program (such as Character Map) and copying the unicode codepoint.

The IconBrush type and editor introduces a new workflow that visually displays the icons available and allows swapping of icons with a mouse click. More information can be found in the Utilities.md

Changes include:

- An Editor Utility Widget for editing UxtIconBrushes (IconBrushEditor)
![image (4).png](https://dev.azure.com/MRDevPlat/ef3b0549-cd24-470e-8e9b-d41a65b99026/_apis/git/repositories/9c84ae5f-0bb9-4823-8c53-0425cd4a6501/pullRequests/344/attachments/image%20(4).png)
- Offline font character querying
- A new IconBrush type to wrap icon properties in a single container and help future proof icon needs with property customization to easily launch the Character Map Viewer
- Integration and documentation with existing content
![image (5).png](https://dev.azure.com/MRDevPlat/ef3b0549-cd24-470e-8e9b-d41a65b99026/_apis/git/repositories/9c84ae5f-0bb9-4823-8c53-0425cd4a6501/pullRequests/344/attachments/image%20(5).png)

Related work items: #485, #493