Merged PR 366: Unifying All Backplate Border Widths
Before this change it was very easy to end up in a situation like this:

![image.png](https://dev.azure.com/MRDevPlat/ef3b0549-cd24-470e-8e9b-d41a65b99026/_apis/git/repositories/9c84ae5f-0bb9-4823-8c53-0425cd4a6501/pullRequests/366/attachments/image.png)

where the border width and radius is different for buttons of different sizes. To hack around this I made a few material instances (MI_ButtonHoloLens2BackPlate**Wide**.uasset) to fix this issue. This was pretty brittle since the user needed to make a new material instance for each backplate size permutation.

Now the backplate material is generated programmatically via the _UUxtBackPlateComponent_ to ensure all borders are uniform no matter the backplate size.

Note, all backplates are now native components. Hence the numerous asset changes.

Related work items: #404