Merged PR 375: Fixing Android Shader Compilation Issues
Adding static to global variables within our shell shaders to allow for assignment. Once this was fixed a new issue was discovered where  FObjectFinder does not pull in the include files referenced by UMaterialExpressionCustom::IncludeFilePaths and fails to compile during cooking. :(

For now FObjectFinder  is commented out and all offending materials are referenced in data (this means our users don't have a nice default state when adding UxtBackPlates, but allows them to compile for Android.) I have a ticket open with Epic on the [best way to proceed](https://udn.unrealengine.com/questions/584444/global-shader-variables-are-const-when-compiling-f.html).

Rendering on Android:
![Android.jpg](https://dev.azure.com/MRDevPlat/ef3b0549-cd24-470e-8e9b-d41a65b99026/_apis/git/repositories/9c84ae5f-0bb9-4823-8c53-0425cd4a6501/pullRequests/375/attachments/Android.jpg)

Related work items: #521