Update UxtGenericManipulatorComponent.cpp
Behaviours of grabbed objects was not as expected when using either MaintainRotationToUser or GravityAlignedMaintainRotationToUser behaviours.

I have altered the code to give the behaviours I expected.

The MaintainRotationToUser behaviour in the MRTK Unity 2.3 shows the same pathologies in ManipulationHandler.cs , but the GravityAlignedMaintainRotationToUser in the same file shows what I believe is the correct behaviour...and that is what I have matched in this UXT version.