Merged PR 384: Loader Open Level
Updating the loader level to perform an open level rather than a stream level. This will allow Framestore's surface magnetism example to function when loaded from the Loader scene.

Because the loader scene is now "unloaded" I've added a BP_ToggleFeaturesPanel to go back to the loader level (Plus, view the current frame stats as requested by Alex Floyd. Note, the frame stats only render in one eye twice, but it is still "functional" and this is something that needs to be fixed by Epic.)

![image.png](https://dev.azure.com/MRDevPlat/ef3b0549-cd24-470e-8e9b-d41a65b99026/_apis/git/repositories/9c84ae5f-0bb9-4823-8c53-0425cd4a6501/pullRequests/384/attachments/image.png)

Related work items: #534