Merged PR 444: Add a camera to MRPawn to get the PlayerCameraManager to track the HMD
The PlayerCameraManager's position was not tracking the HMD position during remoting or packaged for the device. This was due to the view target being the MRPawn, which did not have a camera. By giving the MRPawn a HMD-locked CameraComponent, the PlayerCameraManager's position will now track the HMD position. This didn't occur in PIE as that sets up a separate _HmdCamera_ used for the input simulation.

Also fix a local build break around header ordering and a minor change in `UUxtFunctionLibrary::GetHeadPose` suggested by @<Bernadette Thalhammer 🐬>

Related work items: #580, #581