Merged PR 455: Adding a loading mutex in the loader blueprint.
Loading a level while already loading another level was causing issues. I didn't get any useful info from the logs about the crash, but this "fix" should avoid that scenario from happening altogether. Adding an "is loading" mutex was also the least intrusive fix I could think of. Changes highlighted in blue:

![image.png](https://dev.azure.com/MRDevPlat/ef3b0549-cd24-470e-8e9b-d41a65b99026/_apis/git/repositories/9c84ae5f-0bb9-4823-8c53-0425cd4a6501/pullRequests/455/attachments/image.png)

Luckily, the way we load levels in completely different in the master branch so this shouldn't be an issue in the future.

Related work items: #593