Merged PR 507: Check PR description for disallowed keywords
PullRequest description will now be checked together with Git History as part of the PR validation build (as it will become the commit message when changes are squashed). If the check detects any issues, the PR description will need to be edited and the build should be queued again before merging.
This check is in place regardless of the merge/rebase options selected as the branch policy does not enforce any particular merge method.

Please be careful making last minute edits before merging if you are not planning on running the build again. If you do that, you risk having to go through the painful process of fixing history on master as problems would only be detected in the main CI build.

Related work items: #659