Replace the unreliable Pivot component in TooltipSpawner with a component reference.
The Pivot was created as a owned component of the spawner component, but the component
can get deleted (as happened in our example scene) and there is no check for this case.

Component references, dependencies, required component, etc. seem to be lacking in UE4,
the next best solution is a FComponentReference where the user can set a separete scene
component to be used as pivot. If no pivot component is set, the offset will just be zero.