Merged PR 523: Nativization of BP_TextActor to AUxtTextRenderActor
Moving the implementation of `BP_TextActor` to `AUxtTextRenderActor`. This is a breaking change thus a few assets had to be modified to swap out the blueprint for the native class.

I also took this opportunity to document how we display text in UX Tools and some best practices.

Related work items: #600