Merged PR 545: Close the open hand menu when the selection changes + minor button/slider bugfix
Close the open hand menu when the selection changes

When testing, I noticed the slider on the menu was almost impossible to interact with. This was caused by the collision not being not being property disabled on the slider visuals when the slider starts with its actor collision disabled. Any changes made to component collision while the actor's collision is disabled get overwritten with the previous value when actor collision is re-enabled.

This also affects buttons but wasn't as noticeable as you would hit the box component as you pushed through the visuals collider. Added the fix there too for consistency.

The issue was uncovered by !480, as the previous event model passed down the focus / grab events when the box component was obscured by the visuals.

Related work items: #690