Merged PR 536: Move default button configuration out of DefaultUXTools.ini
This is to fix #688, which was introduced in !477 and repros in every pipeline artifact I have tested since those changes.

The root cause of this is that the build pipeline is not cooking the content referenced from DefaultUXTools.ini, which was being used for some default button configuration.

The fix for this is straightforward, it's simply moving the default button configuration from _DefaultUXTools.ini_ to the button constructors. This way the cooker is aware of them regardless of how the environment / pipeline is set up.

If you want to test out before / after packages you can get them here:
* [Broken build](https://dev.azure.com/MRDevPlat/DevPlat/_build/results?buildId=1738&view=results)
* [Fixed build](https://dev.azure.com/MRDevPlat/DevPlat/_build/results?buildId=1742&view=results)

Related work items: #688