Merged PR 559: Large hand menu closes unexpectedly
Update the _lock on grab_ hand menu to no longer re-attach to the hand when world locked. The lack of "require eye gaze" trigger was causing the menu disappear when the hand temporarily registered in the flat palm pose when moving around.

The _lock on drop_ menu keeps the re-attach logic as it was very easy to "lose" the menu when a hand registered as a flat palm when outside of the device's viewport.

Added a note to #692 add back re-attach to the _lock on grab_ menu when that feature is implemented

Related work items: #718