Merged PR 565: Back Plate Rotation Fix
"Rotating" the back plate so that it's local up is the z-axis and not the y-axis. At the moment this was accounted for by rotating the mesh component relative to it's parent. But, Unreal doesn't account for parent child rotations when applying parent child scales, so you could run into situations where scaling the parent would not scale the child back plate appropriately.

The ideal fix would be to re-import the back plate mesh with the correct orientation. But, the back plate shader makes copious assumptions about it's local up axis. Rather than try to resolve these assumptions (and break future compatibility) we rotate the vertices in the vertex shader , rotate the bounds, and rotate the collider.

After making these changes I had to fix-up content that was scaled as "height x width x depth" to "depth x width x height."

This change allows the user to perform actions like this now:

![ScaleButton.gif](https://dev.azure.com/MRDevPlat/ef3b0549-cd24-470e-8e9b-d41a65b99026/_apis/git/repositories/9c84ae5f-0bb9-4823-8c53-0425cd4a6501/pullRequests/565/attachments/ScaleButton.gif)

![Scale.gif](https://dev.azure.com/MRDevPlat/ef3b0549-cd24-470e-8e9b-d41a65b99026/_apis/git/repositories/9c84ae5f-0bb9-4823-8c53-0425cd4a6501/pullRequests/565/attachments/Scale.gif)

Related work items: #724, #726