Merged PR 574: Bounds Control Affordance Colors When Grabbed
This PR addresses bug #641 by changing the bounds control "cage" to blue and solid when grabbed. It also adjust the affordances to be opaque and unlit. It also ensures only one affordance is visible (and stays visible) when grabbed. (Also, added the correct audio for bounds controls and removed unused materials.)

![BoundsControl.gif](https://dev.azure.com/MRDevPlat/ef3b0549-cd24-470e-8e9b-d41a65b99026/_apis/git/repositories/9c84ae5f-0bb9-4823-8c53-0425cd4a6501/pullRequests/574/attachments/BoundsControl.gif)

Speaking with @<Lukas Tönne> offline he is also working on the bounds control. So I tried not to get too invasive with changes. Some other things to address is to see if we can eliminate the need for the blueprints to add a bounds control "cage." The cage should be spawned by the bounds control component or actor. The material swap should also happen outside the BP, and audio should spawn outside the BP.

Related work items: #641, #750