Merged PR 617: Add gaze requirement to PalmUpConstraint and Hand Menu
* Add gaze requirement to `UUxtPalmUpConstraintComponent`
* Enable _Require Gaze_  on hand menu examples
* Re-enable re-attach logic on the world-lock on grab hand menu (!559)
* Update docs and build call steps

MRTK-Unity uses thumb and index fingers when calculating the center of the gaze activation zone for the ulnar side of the hand. Instead we are using the little finger metacarpal for this as every other case moves the center towards the desired zone. I've logged a bug on MRTK-Unity [here](https://github.com/microsoft/MixedRealityToolkit-Unity/issues/8896).

Related work items: #692