Merged PR 623: Fix one handed only manipulation not responding if grabbed by both hands
Currently if you set a generic manipulator to use only one handed manipulation and grab it with both hands it will no longer respond to interactions. Fixing this at the GrabTargetComponent level gives a better user experience as we can prevent the second grab entirely instead of just ignoring the second pointer, which is what MRTK-Unity does.

* Move the _ManipulationModes_ setting up to the `UxtGrabTargetComponent` and rename it to _GrabModes_
* Update the `UxtGrabTargetComponent` to:
  * Only accept grab pointers up to the _GrabMode_ limit
  * Trigger grab/release events when it's grab mode has been satisfied and not for every grab (e.g. when using only two handed grabs, a grab event is only triggered when both hands are grabbing the object)
* Improve the `UxtGrabTargetComponent` tests
  * Clean up exiting interaction tests
  * Introduce tests for focus/grab events
  * Introduce tests for grab mode settings
* Up the limit on when to use unity builds for the tests, as they cannot compile in a unity build

Related work items: #827