Fix missing left MotionController on Android devices.
Creating the UMotionControllerComponent in the ctor of the UxtHandInteractionActor
means that the class default object (CDO) also has a motion controller component.

The way the default XRTrackingSystemBase implementation works means that it looks for
all instances of the UMotionControllerComponent, then picks the first one that matches
the requested motion source (i.e. hand). For the left hand that will be the actor's CDO
motion controller, which of course is never tracked or providing data.

Creating the motion controller components in BeginPlay means they only exist at runtime
and the HMD will pick valid instances for retrieving data.