Merged PR 644: Disabling unfinished "Mask Logic" in Scrolling Object Collection
I was able to reproduce #898 thankfully pretty quickly. It looks like a race condition between the scrolling object collection the text renderer component setting constant buffer values. The scrolling object collection never finished support for masking, so this feature was half baked (and not the way I would recommend doing it, ie a new instance for every material).  So, for a "quick fix" I've disable the masking logic until we implement the "real solution."

![image.png](https://dev.azure.com/MRDevPlat/ef3b0549-cd24-470e-8e9b-d41a65b99026/_apis/git/repositories/9c84ae5f-0bb9-4823-8c53-0425cd4a6501/pullRequests/644/attachments/image.png)

Images of corruption:

![Bug.JPG](https://dev.azure.com/MRDevPlat/ef3b0549-cd24-470e-8e9b-d41a65b99026/_apis/git/repositories/9c84ae5f-0bb9-4823-8c53-0425cd4a6501/pullRequests/644/attachments/Bug.JPG)
![Bug2.JPG](https://dev.azure.com/MRDevPlat/ef3b0549-cd24-470e-8e9b-d41a65b99026/_apis/git/repositories/9c84ae5f-0bb9-4823-8c53-0425cd4a6501/pullRequests/644/attachments/Bug2.JPG)
![Bad.JPG](https://dev.azure.com/MRDevPlat/ef3b0549-cd24-470e-8e9b-d41a65b99026/_apis/git/repositories/9c84ae5f-0bb9-4823-8c53-0425cd4a6501/pullRequests/644/attachments/Bad.JPG)
![Good.JPG](https://dev.azure.com/MRDevPlat/ef3b0549-cd24-470e-8e9b-d41a65b99026/_apis/git/repositories/9c84ae5f-0bb9-4823-8c53-0425cd4a6501/pullRequests/644/attachments/Good.JPG)

Related work items: #898