Merged PR 677: Fix-up example assets to allow UXToolsGame enable Blueprint nativization
## Summary
No functionality changes in this PR, just fixing up of some assets to allow them to be nativized.

Note: Enabling nativization will cause packaging to fail on 4.26, this was fixed in 4.26.1.

## Changes
The only major change is to the assets based on Near Menus, they had to be re-worked to not inherit child actors to work around a crash caused by https://issues.unrealengine.com/issue/UE-87474.
This adds a new `BP_NearMenuBase` which replaces `BP_NearMenu2x1` as the base for the menus. It includes all the functionality of `BP_NearMenu2x1` but without the buttons. Buttons are now added to each menu as needed.

The rest of the assets are just removing and re-adding `UxtBackPlateComponents` and `UxtGenericManipulatorComponents` to fix up missing data that works in blueprint interpreter but not the nativization path.

### Detailed changes
_Replace_ here means remove and re-add the component.
- `BP_CoffeeCup`: Replace manipulator
- `BP_Slate2D`: Replace manipulator + backplate
- `BP_EarthCore`: Replace manipulator
- `BP_DescriptionPanel`: Replace backplate
- `BP_NearMenuBase`: New base class for near menus, includes all the functionality previously provided by `BP_NearMenu2x1`
- `BP_ToggleFeaturesPanel`: Replace base class with `BP_NearMenuBase`
- `BP_HandMenuSlider`: Adds variable to prevent it being marked as data-only to work around https://issues.unrealengine.com/issue/UE-44213
- `BP_LargeHandMenu_LockOnDrop`: Replace manipulator, backplate + add face user constraint
- `BP_LargeHandMenu_LockOnGrab`: Replace manipulator, backplate + add face user constraint
- `BP_MediumHandMenu`: Replace backplate
- `BP_SelectorMenu`: Replace backplate
- `BP_SmallHandMenu`: Replace backplate
- `BP_NearMenu2x1`: Replace base class with `BP_NearMenuBase`
- `BP_NearMenu3x1`: Replace base class with `BP_NearMenuBase`
- `BP_NearMenu3x2`: Replace base class with `BP_NearMenuBase`
- `BP_NearMenu3x3`: Replace base class with `BP_NearMenuBase`
- `BP_NearMenu4x1`: Replace base class with `BP_NearMenuBase`
- `BP_NearMenu4x2`: Replace base class with `BP_NearMenuBase`
- `BP_ScrollingObjectCollection_SortByButtonLabel`: Replace backplate
- `BP_ScrollingObjectCollection`: Replace backplate

Related work items: #614