Merged PR 680: Remove location offsets from Slate2D's bounded meshes
Removes the location offsets of the meshes in `BP_Slate2D.uasset`, translating the instance in the example scene instead.

`Bounds` are calculated automatically in `BoundsOverride`'s local space and its centre is the pivot for Bounds Control rotations. Therefore, having location offsets with respect to that space ends up in weird offsets for affordances when transforms start to add up.