Move XRSimulation parts out of the UxtDefaultHandTrackerSubsystem.
Simulated hand state was persisting between game sessions in the editor,
which is undesirable. The reason is that UxtDefaultHandTrackerSubsystem is
an engine subsystem which means it persists as long as the editor is running
and so does the simulation state.

Resetting the simulation state in the PostLogin/Logout events does not work,
because these are actually called for every map change. The hand state would
be reset every time a map is loaded, which is also not desirable. (That was
the reason the separate simulation state was introduced in the first place,
so it could persist across map loads).

A LocalPlayerSubsystem solves the issue because the local player is persistent
across maps, but gets destroyed and recreated for each game session.

An added benefit of this change is that the XRSimulation-related functionality
is moved out of the UxtDefaultHandTrackerSubsystem, so it becomes easier to
remove again once the HMD class becomes a viable home for it in the future.