Merged PR 762: Enable XRSimulation to be built on Hololens and Android
Packaging UXToolsExamples will fail to compile on HoloLens and Android due to missing object files from UXTools' XRSimulation module, which is only set to build for Win64.

The UXToolsInput module build for all three platforms and has a dependency on XRSimulation, causing the packaging issue for non-Win64 platforms. I'm not sure why packaging the game works with the current setup but packaging plugins doesn't.

The safest fix for 0.12 is to enable compilation of XRSimulation for HoloLens and Android. It might be worth looking into abstracting XRSimulation away from UXToolsInput if we don't get in-engine XR input sim in the future.

Related work items: #1078