Merged PR 792: Have a hand menu type selected by default in hand menu map
It may not be immediately obvious that the different menu types are buttons you can click on. I changed it so the small menu is selected by default, which makes it more obvious and makes the example showcase one of the menus by default instead of nothing.

Related work items: #1095