Merged PR 754: Create Bounds Control mesh automatically
- Gets rid of the artificial `CollisionBox` added previously to prevent
far rays from passing through the box. This is now achieved with the
`BoundingBoxComponent`.
- Gets rid of the `BoundControl` and related stuff in
`BP_CoffeeCup.uasset` and `BP_Slate2D.uasset`. With the previous
approach, users needed to replicate this behaviour manually, which would
lead to issues when they don't (GitHub issue #25). It's handled
automatically from C++ now.
- Collision box of `SM_BoundsControl.uasset` now matches the mesh
exactly. It used to extend slightly past the mesh itself, which is
desirable in affordances so it's easier to target them, but not in the
case of the bounding box mesh, which only needs to block rays passing
through it without interfering in affordances' collisions.