Fix input simulation when loading maps in multiplayer testing.
This was broken because the subsystem creating simulation actors was
only checking if a player controller is local before assigning the
actor to it. In a multiplayer test with several client windows on the
same machine ALL player controllers are local, so on the second login
the first client that logs in has all other clients' simulation actors
also assigned to it.

The subsytem is a LocalPlayerSubsystem, so it should just check if
a player controller is for its own LocalPlayer.