Merged PR 938: Add a list of allowed absolute URLs in Docs
The absolute URL to Release Notes for major 0.12.0 release causes docs validation to fail. I think it would be good to include this link, so I added a list of exceptions. A warning should be shown on the main build results page for each URL encountered to make it less likely to forget that we need to remove those links when no longer needed.

Related work items: #1018