Avoid loading order problem when using anim blueprint class.
Loading the animation blueprint for XRSimulation will require the
PropertyAccessEditor modular feature to exist, but XRSimulation settings
are created in PostConfigInit and the feature is only registered in the
later PreDefault stage.

To avoid loading the blueprint class, use a SoftClassPtr which takes
only a reference path to initialize. The asset can then be loaded lazily
when the XRSim hands are actually created.